extended browserify sources to be able to analyse both external and internal dependencies
external dependencies are dependencies that have been published as an npm,
but browserify bundles can also contain internal modules that will never be published as an npm.
this extension simply adds the possibility to analyse internal dependencies with browserify-licenses as well as external dependencies,
while keeping existing analysis for external dependencies working as before.